feat: add SilencePlayerListener to handle vanish player interactions
