+++ gradle.properties
@@ -1,4 +1,4 @@
 kotlin.code.style=official
 kotlin.stdlib.default.dependency=false
 org.gradle.parallel=true
-version=1.21.10-1.0.0-SNAPSHOT
+version=1.21.10-1.0.1-SNAPSHOT
