+++ surf-vanish-paper/src/main/kotlin/dev/slne/surf/vanish/paper/PaperMain.kt
@@ -7,18 +7,18 @@ import dev.slne.surf.vanish.paper.command.vanishCommand
 import dev.slne.surf.vanish.paper.config.VanishConfiguration
 import dev.slne.surf.vanish.paper.listener.ConnectionListener
 import dev.slne.surf.vanish.paper.listener.HotkeyListener
-import dev.slne.surf.vanish.paper.listener.SilenceListener
+import dev.slne.surf.vanish.paper.listener.SilencePlayerListener
 import org.bukkit.plugin.java.JavaPlugin
 
 val plugin get() = JavaPlugin.getPlugin(PaperMain::class.java)
 
 class PaperMain : SuspendingJavaPlugin() {
     override fun onEnable() {
-        vanishCommand()
-
         ConnectionListener.register()
-        SilenceListener.register()
+        SilencePlayerListener.register()
         PacketEvents.getAPI().eventManager.registerListener(HotkeyListener)
+
+        vanishCommand()
     }
 
     override fun onDisable() {
